commit_83660 ------------------------------------------------------------------------ r83660 | vboxsync | 2020-04-10 02:06:53 +0800 (Fri, 10 Apr 2020) | 1 line
IPRT/log-vbox.cpp: Use timestamped log files in GAs too, not just VBox.log.  Unlike on the host, the files are prefixed 'VBoxGAs-' so they are easier to find and delete.  Not printing any header as I'm unsure about the amount of stack space we can use when called, playing it safe for now.  Also reduced the stack space used on the host side by combining the executable path buffer with szBuf. bugref:9670
------------------------------------------------------------------------