commit_83958 ------------------------------------------------------------------------ r83958 | vboxsync | 2020-04-24 04:24:40 +0800 (Fri, 24 Apr 2020) | 1 line
VBoxServiceVMInfo.cpp: Two NT4 fixes.  First, the NT4 SP1 workaround for the uninitialized variable in GetTcpipInterfaceList wasn't active due to the lowercased x in RT_ARCH_X86.  Second, we were walking a IP_ADAPTER_INFO list even when GetAdaptersInfo wasn't actually called or failed with ERROR_NO_DATA. Set the pAdpInfo and cbAdpInfo variables to NULL and 0 in that case, as the code expects it to be (see NULL checks before the strcmp loop we crashed, or in one instance point got stuck, in). Also, initialize the buffers with zeros.  bugref:8489
------------------------------------------------------------------------