commit_84314 ------------------------------------------------------------------------ r84314 | vboxsync | 2020-05-15 13:13:53 +0800 (Fri, 15 May 2020) | 4 lines
AMD IOMMU: bugref:9654 DSFX, DSCX, DSSX revision numbers. Not sure we need them, but one of the BKDG (not all) has PCI config.
registers that report Marc bits via DSSX. Not yet sure we need to implement all those PCI config. registers as its not mentioned
anywhere in the actual spec. but only in the BKDG.

------------------------------------------------------------------------