commit_84326 ------------------------------------------------------------------------ r84326 | vboxsync | 2020-05-18 11:23:00 +0800 (Mon, 18 May 2020) | 8 lines
OCI: ICloudClient::listSourceBootVolumes - like listBootVolumes but
filters out boot volumes that are not suitable as a source for a new
cloud machine (dying/dead, attached to an instace, already used by a
cloud machine).

While here declare listSourceInstances with the similar intended
semantics, but provide only a stub implementation for now.

------------------------------------------------------------------------