commit_84849 ------------------------------------------------------------------------ r84849 | vboxsync | 2020-06-16 22:23:39 +0800 (Tue, 16 Jun 2020) | 1 line
DevVGA-SVGA3d: Jettison the titlebar and border of the window we pass to CreateDeviceEx, as things goes sideways if we don't have any pixels in the client area.  The titlebar size and stuff changes with the subsystem/os version number in the PE header, so when the new linker started putting 6.00 there instead of 5.02, things stopped working on a bunch of W10 systems. bugref:8489
------------------------------------------------------------------------