commit_84968 ------------------------------------------------------------------------ r84968 | vboxsync | 2020-06-26 19:51:54 +0800 (Fri, 26 Jun 2020) | 2 lines
Installer/linux: Bring DEB build/package to same functionality as the recent changes for RPM (needing WINE less often, due to not building the CHM docs). Additionally a very crude (manually specified package deps are still present) way of suppressing the build of VBoxSDL, as another measure to make the functionality as identical as possible.

------------------------------------------------------------------------