commit_85069 ------------------------------------------------------------------------ r85069 | vboxsync | 2020-07-06 21:01:22 +0800 (Mon, 06 Jul 2020) | 1 line
iprt/cdefs.h: Map RT_NO_THROW_PROTO/DEF onto noexcept when possible as throw() is being redefined in c++17 to involve terminate.  Also, Clang (11 at least) get upset when we use __nothrow__ on the prototype but not on the definition and we'd need a DECL_NO_THROW for putthe the __nothrow__ attribute in the right place on a definition, which is too much work right now.  bugref:9790
------------------------------------------------------------------------