commit_85267 ------------------------------------------------------------------------ r85267 | vboxsync | 2020-07-12 09:09:57 +0800 (Sun, 12 Jul 2020) | 1 line
Main/NetIf-generic.cpp: Made NetIfAdpCtl() return failure when the exit code is non-zero or it terminated abnormally (judging from the error messages, this is what the callers expects).  Tried to untangle status code mess in NetIfCreateHostOnlyNetworkInterface(), adding missing error handling for popen() and adjusting some of the copy&paste error messages a little.  NetIfRemoveHostOnlyNetworkInterface() should not call i_notifyComplete on an uninitialized progress object, and the return of Progress::init() is not a vbox status code that this function can return unconverted.  createObject can fail. [build fix] bugref:9790
------------------------------------------------------------------------