commit_86184 ------------------------------------------------------------------------ r86184 | vboxsync | 2020-09-20 20:08:14 +0800 (Sun, 20 Sep 2020) | 1 line
Debugger/DBGCRemoteKd: Implement access to the internal VBox debugger. The protocol doesn't support something similar to GDBs monitor command which translates to a Rcmd packet unfortunately. However, KD supports a driver querying the debugger for some string input via DbgPrompt() which translates to DEBUG_IO packets being exchanged for setting the prompt and reading the response. This needs to be initiaed from our stub though, which is done by abusing the CauseBugCheck request is abused to drop into the VBoxDbg prompt. The request can be issued with WinDbgs .crash command
------------------------------------------------------------------------