commit_86495 ------------------------------------------------------------------------ r86495 | vboxsync | 2020-10-08 21:04:26 +0800 (Thu, 08 Oct 2020) | 1 line
VBoxSVC/HostPowerServiceLinux: Use RTThreadPoke on the notification thread if it doesn't seem to respond to dbus_connection_close() after 10 ms.  Also, grab an extra reference on the connection so both the thread and the destruction have one to drop, otherwise we'd be in trouble if the other side terminates the connection and the thread exits early. bugref:9841
------------------------------------------------------------------------