commit_86681 ------------------------------------------------------------------------ r86681 | vboxsync | 2020-10-23 07:36:16 +0800 (Fri, 23 Oct 2020) | 9 lines
IPRT/rest: bugref:9167 - Add release logging for REST operations.
Request are logged at level 5, and responses at level 7, so they are
not on by default.  The format is a bit ad-hoc, but all the essential
information should be there.  It's mostly done in base classes except
for request bodies (it can probably be done in the base class too, but
would require more runtime logic, while here we can only generate them
as needed).

------------------------------------------------------------------------