commit_87372 ------------------------------------------------------------------------ r87372 | vboxsync | 2021-01-22 23:01:51 +0800 (Fri, 22 Jan 2021) | 1 line
VMM/HMR0A.asm: Baked the XMM saving and loading into the hmR0SvmVmRunTemplate and ditched the hmR0SVMRunWrapXMM wrapper.  It's not entirely perfect yet as the decision whether we need to do anything with the SSE state and how we do it is take each time we switch, but since the most important piece of info is in CPUM it's a little bit hard to do from HM atm.
------------------------------------------------------------------------