commit_87703 ------------------------------------------------------------------------ r87703 | vboxsync | 2021-02-11 06:12:40 +0800 (Thu, 11 Feb 2021) | 2 lines
Installer/solaris: Additional amendment to vbox-ips.mog to remove dependency on python 2.6 since VBoxPython.so is linked with libpython2.6.so.1.0 on Solaris 11 systems up to 11.3 SRU19 inclusive.  Also change vboxshell.py to explicity invoke python2.7 which has always been available on Solaris 11 and is delivered in S10 as part of freely available patches. bugref:5999

------------------------------------------------------------------------