commit_88039 ------------------------------------------------------------------------ r88039 | vboxsync | 2021-03-09 20:28:53 +0800 (Tue, 09 Mar 2021) | 9 lines
VBoxManage/modifyvm: Don't try to always helpfully enable a NIC when
changing its attachment, only do that if it's not already enabled.

For a VM in a saved state the attachment can be changed, but the NIC
status cannot, so it was impossible to change NIC attachments from the
command line in that case.  Of course, the setter also should not
freak out about a no-op request to enable an already enabled NIC and
that should be fixed, but better be twice shy...

------------------------------------------------------------------------