commit_88079 ------------------------------------------------------------------------ r88079 | vboxsync | 2021-03-11 04:48:09 +0800 (Thu, 11 Mar 2021) | 7 lines
VBoxManage: When setting the host interface for bridging or hostonly
attachments do a sanity check on it.  Make sure IHost knows about it
and that the interface is of correct type.  Warn the user if that's
not the case, but the warning is non-fatal.  The API setters don't do
this, so provide at least some protection from typos and mistakes.
bugref:9966.

------------------------------------------------------------------------