commit_88137 ------------------------------------------------------------------------ r88137 | vboxsync | 2021-03-16 20:22:05 +0800 (Tue, 16 Mar 2021) | 1 line
DevHDA: Made the DMA timer scheduling heuristics more flexible and ditched the alternative approaches. Changed the DMA engine to load the whole BDL and not reload BDLEs as we work thru them.  Ditched the HDABDLE structure and some other DMA related stuff (FIFO buffer is on the todo list). bugref:9890
------------------------------------------------------------------------