commit_88199 ------------------------------------------------------------------------ r88199 | vboxsync | 2021-03-19 04:09:11 +0800 (Fri, 19 Mar 2021) | 2 lines
Main/Makefile.kmk: Downgrade gcc 7+ optimizations a little because they otherwise create symbols which are longer than dtrace can deal with on Solaris 11.4. The limit is reportedly 148 bytes. Not showing up on older Solaris because we use gcc without these aggressive optimizations which create those extra functions due to code reordering to streamline the likely taken path. No idea why this does not happen on Linux with new gcc and our own dtracei, because the mangling should be the same.

------------------------------------------------------------------------