commit_88269 ------------------------------------------------------------------------ r88269 | vboxsync | 2021-03-24 19:45:54 +0800 (Wed, 24 Mar 2021) | 1 line
Audio: Made sure PDMAUDIOPCMPROPS is initialized using a helper function or the initializer macro, and that vital changes are made using setting helper functions.  There are now two derived fields (frame size and shift count) that must be maintained, so this was the sanest way of doing it.  Added a raw flag to PDMAUDIOPCMPROPS for VRDE/VRDP, since it wants the raw mixer content and we need a way of expressing this (PDMAUDIOSTREAMLAYOUT isn't the right place).  The mixer buffers now uses PDMAUDIOPCMPROPS rather than the weird 32-bit format contraption for picking conversion functions.  Simplify the drvAudioStreamPlay code by eliminating the PDMAUDIOSTREAMLAYOUT_RAW special case.  bugref:9890
------------------------------------------------------------------------