commit_88376 ------------------------------------------------------------------------ r88376 | vboxsync | 2021-04-07 04:35:18 +0800 (Wed, 07 Apr 2021) | 1 line
DrvHostAudioDSound.cpp: Ripped out the extra ~600ms buffering.  When asked to DRAIN the backend, try switch the buffer to non-looping mode so that it will stop when it's done (may need to clear the rest, not sure).  Don't RESUME disabled streams - this worked because we used to fill the buffer with silence upon disabling a stream (no longer true with new draining code).  bugref:9890
------------------------------------------------------------------------