commit_88433 ------------------------------------------------------------------------ r88433 | vboxsync | 2021-04-09 20:55:19 +0800 (Fri, 09 Apr 2021) | 1 line
Audio: Eliminated the DrvAudio mixing buffers for output streams on devices without their own mixer. Changed the prebuffering to not default to the whole backend buffer size, but only 2/3 of it, so that there is room for a bit of incoming data from the device once we start playing.  We don't want to have that gather in the device mixer or internal DMA buffers.  This code needs some more testing and work, only tested on linux against ALSA.  bugref:9890
------------------------------------------------------------------------