commit_88853 ------------------------------------------------------------------------ r88853 | vboxsync | 2021-05-04 16:47:13 +0800 (Tue, 04 May 2021) | 1 line
Audio: Reworking PDMIHOSTAUDIOPORT::pfnNotifyDevicesChanged a little, from now on it will only trigger re-enumeration (darwin still needs some adjusting).  Use a timer to sligtly delay the enumeration and move it out of the re-init stream code, so that we don't redo it too frequently and that we do it even when there are active streams around.  Made DrvHostAudioWasApi call pfnNotifyDevicesChanged.  Reduced the DSound notification client code a little while adding pfnNotifyDeviceChanged calls to it.   bugref:9890
------------------------------------------------------------------------