commit_88991 ------------------------------------------------------------------------ r88991 | vboxsync | 2021-05-12 08:46:35 +0800 (Wed, 12 May 2021) | 1 line
Audio: Worked over draining, starting with the internal DMA buffer (instead of just the pre-buffer and backend buffer) and using the async I/O thread to keep calling PDMIAUDIOCONNECTOR::pfnStreamIterate and PDMIHOSTAUDIO::pfnStreamPlay (NULL buffer) every so often till the draining is done.  Also put a rough deadline on the draining.  The PDMAUDIOSTREAMCMD_DISABLE is now defined to stop playback/capturing immediately, even when already draining (if possible).  This gets rid of the timers in DrvAudio and windows backends.  DrvAudio no longer issue an DISABLE command at the end of the drain, it assumes the backend does that internally itself.  After issuing PDMAUDIOSTREAMCMD_DRAIN the client (be it mixer or drvaudio) will not provide any more data for the buffers via pfnStreamPlay.  Only tested windows, needs to re-test all platforms.  bugref:9890
------------------------------------------------------------------------