commit_89028 ------------------------------------------------------------------------ r89028 | vboxsync | 2021-05-13 14:15:39 +0800 (Thu, 13 May 2021) | 4 lines
Intel IOMMU: bugref:9967 Handle cases when the invalidation complete event interrupt is masked but unmasked later.
I'm not entire sure if the MSI message can be cached at the time the event was made pending or if the active IEADDR_REG and IEDATA_REG values
can be used to reconstruct the MSI. Using the latter for now.

------------------------------------------------------------------------