commit_89907 ------------------------------------------------------------------------ r89907 | vboxsync | 2021-06-25 07:23:21 +0800 (Fri, 25 Jun 2021) | 7 lines
FE/VBoxHeadless: Enable signal handlers on all unix platforms.  Note
in the function's comment for showProgress() that we don't check for
signals in b/c the power up/down operations are not cancelable anyway.
Establish the signal handlers earlier.  Add a todo item to convert any
temporary changes we do to the settings to use RAII so that they are
undone automatically.  bugref:8161.

------------------------------------------------------------------------