commit_89917 ------------------------------------------------------------------------ r89917 | vboxsync | 2021-06-26 02:35:48 +0800 (Sat, 26 Jun 2021) | 4 lines
docbook2latex.xsl: Fix incorrect formatting (had too much line spacing and also was justified) of the last line of a refentry icommand synopsis. Use standard font (not monospace) for the square brackets, alternatives and so on. Use ellipses and tweak line breaking to happen at a defined places with alternatives. More compact and easier to read. Made the XSLT a little simpler with args and groups handling, and resolved a couple of quirks with missing or extra spaces when nesting them.

For several man_VBoxManage-*.xml files: Eliminate redundant attributes (choice="opt" is default) to make the files easier to read. Unified DOCTYPE and applied some small cleanups here and there.

------------------------------------------------------------------------