commit_90234 ------------------------------------------------------------------------ r90234 | vboxsync | 2021-07-17 06:28:08 +0800 (Sat, 17 Jul 2021) | 6 lines
VBoxManage: Give showProgress() a flags argument so that it can poll
the progress quietly.  While here, teach it to print the progress
description (alone, or before the running progress count).  No
existing callers should be affected, though the couple of callers that
use g_fDetailedProgress should be converted to the the flags instead.

------------------------------------------------------------------------