commit_91170 ------------------------------------------------------------------------ r91170 | vboxsync | 2021-09-09 03:21:10 +0800 (Thu, 09 Sep 2021) | 8 lines
VBoxManage: With SHOW_PROGRESS_NONE don't print "progress state" when
the async operation completes with an error.  bugref:10065.

That's a bit ambiguously phrased too, as it's not the state of the
progress (object itself) that is being reported here, but the state of
the async operation progress this progress object is tracking.  But
that's not the yak I want to shave.

------------------------------------------------------------------------