commit_91338 ------------------------------------------------------------------------ r91338 | vboxsync | 2021-09-23 08:21:53 +0800 (Thu, 23 Sep 2021) | 16 lines
VBoxManageCloudMachine: bugref:10065.  Refactor the code some more.
Extend HandlerArg to carry the stuff we need instead of passing around
a bunch of arguments.

While here accept machie spec after "machine" too.  In that position
an explicit --id/--name option is required for the obvious reason.
We will also check for these options and machine spec as a plain words
argument after the command word, so user can use either of:

  VBoxManage cloud ... machine --name foo start
  VBoxManage cloud ... machine start --name foo
  VBoxManage cloud ... machine start foo

where the first form is easier to wrap into a convenience shell
function.

------------------------------------------------------------------------