commit_91919 ------------------------------------------------------------------------ r91919 | vboxsync | 2021-10-21 09:10:52 +0800 (Thu, 21 Oct 2021) | 9 lines
IPRT: Make RTNetStrToIPv4Cidr() and RTNetStrToIPv4Cidr() accept zero
prefix, mostly for the benefit of 0.0.0.0/0 and ::/0 respectively.

While here fix RTNetStrToIPv6Cidr to do stricter checks on the prefix.
When the number is too big RTStrToUInt8Ex returns a warning, not an
error.  bugref:9330.

(should have been part of r147723).

------------------------------------------------------------------------