commit_92064 ------------------------------------------------------------------------ r92064 | vboxsync | 2021-10-26 15:42:51 +0800 (Tue, 26 Oct 2021) | 1 line
Audio/Validation Kit: On some testboxes draining never seems to finish and thus is getting aborted, no clue why. The test result in the end still could be correct, although the actual draining problem needs to be investigated further. So don't make this (and the stream state check below) an error for now and just warn about it. bugref:10008
------------------------------------------------------------------------