commit_92409 ------------------------------------------------------------------------ r92409 | vboxsync | 2021-11-13 06:17:42 +0800 (Sat, 13 Nov 2021) | 1 line
VMM/PDM,PGM: Added PGMR3IsNemModeEnabled and made PDM not put devices in ring-0 when the simplified memory mode is active as it is implied that we won't be running anything from ring-0 (at least not on windows and mac, I think). PDM must use RTMemAllocZ instead of MMR3HeapAllocZEx for allocating device instance data to get proper alignment. bugref:10122
------------------------------------------------------------------------