commit_92872 ------------------------------------------------------------------------ r92872 | vboxsync | 2021-12-11 08:29:18 +0800 (Sat, 11 Dec 2021) | 1 line
IPRT/timer-r0drv-nt.cpp: Corrected the CPU parameter for rtMpNtSetTargetProcessorDpc in the fSpecificCpu case, we were passing the idCpu instead of the set index.  Skip setting uNtStartTime and uNtDueTime when starting a one-shot timer.  Skip dequeuing DPCs when stopping high-res timers that aren't targetting a specific CPU or are omni timers, as we don't be using any DPCs then.
------------------------------------------------------------------------