commit_93118 ------------------------------------------------------------------------ r93118 | vboxsync | 2022-01-04 09:41:47 +0800 (Tue, 04 Jan 2022) | 1 line
IPRT/RTIsoMaker: Added an --name-setup-from-import option to better handle importing old ISOs without joliet present.  We get into trouble when adding our files to the ISO in both primary iso and joliet space, since the guest might be inclined to show the joliet view (which is missing the files from the imported ISO).
------------------------------------------------------------------------