commit_93119 ------------------------------------------------------------------------ r93119 | vboxsync | 2022-01-04 10:11:01 +0800 (Tue, 04 Jan 2022) | 1 line
VMM/HMSVMR0: Must deal with VINF_IEM_RAISED_XCPT in the string I/O case of an IOIO exit.  Debug builds would guru if the guest was using invalid memory and IEM was forced to raise a #PF or #GP. Reordered fCtxChanged/rcStrict updating in similar code paths, placing the rcStrict after the fCtxChange since that's closer to the return point and might help with register liveness.
------------------------------------------------------------------------