commit_93127 ------------------------------------------------------------------------ r93127 | vboxsync | 2022-01-05 09:10:34 +0800 (Wed, 05 Jan 2022) | 1 line
Main/UnattendedOs2Install: Fixed most of the basic issues, only Netscape left. Added reboot after installing GENGRADD because the default VGA driver is horrendously sluggish.  Use the --as-zero option to deal with the speciall CID and NVDM/2 status codes we're getting on success.  Use the --file-to-backdoor option in failure cases when we've got useful logs.
------------------------------------------------------------------------