commit_93468 ------------------------------------------------------------------------ r93468 | vboxsync | 2022-01-28 05:17:12 +0800 (Fri, 28 Jan 2022) | 1 line
VBoxDbg,VMM/STAM,Main: Converted VBoxDbg to use the VMM function table, extending the STAMR3Enum to include the unit string to avoid needing to call STAMR3GetUnit a lot.  The latter would better fit with the XML based viewer, as it only gets the string version of the unit.  Had to adjust a user of STAMR3Enum in Main. bugref:10074
------------------------------------------------------------------------