commit_93480 ------------------------------------------------------------------------ r93480 | vboxsync | 2022-01-29 00:09:52 +0800 (Sat, 29 Jan 2022) | 8 lines
Main/Appliance: Allow users to specify a different storage controller
and/or controller port for hard disks when importing a VM. bugref:5027

'VBoxManage import foo.ova -n' has always presented a --controller
option for hard disks but the code to implement this had never been
implemented.  This changeset adds the --controller functionality and
also includes a --port option for changing the controller port as well.

------------------------------------------------------------------------