commit_93631 ------------------------------------------------------------------------ r93631 | vboxsync | 2022-02-07 08:45:08 +0800 (Mon, 07 Feb 2022) | 1 line
VMM/PDMNetShaper: Don't start the unchoke thread unless there are are one or more bandwidth groups configured.  Make the unchoke thread wait on an event sempahore so it can be woken up when needed and stop wasting cpu cycles and battery on lots of unnecessary wakups.  Use a timer (real time clock) to wake up the unchoke timer when needed.  bugref:10093 bugref:5582
------------------------------------------------------------------------