commit_93849 ------------------------------------------------------------------------ r93849 | vboxsync | 2022-02-19 00:55:24 +0800 (Sat, 19 Feb 2022) | 1 line
Validation Kit/unit tests: Make sure to mount the Validation Kit .ISO so that TxS has the chance to update itself. This is necessary as a lot of our test VMs nowadays have a very old TxS installed which don't understand commands like uploading files to the guest. Uploading files is needed for this test driver, however. bugref:10195
------------------------------------------------------------------------